Finish support for typing geminated l directly.
The chosen solution has three parts. For unicode capable engines, we
define shorthands ·l and ·L to output the interpunct and the
coressponding l with appropriate kerning. For non-unicode engines, when
working with LaTeX we redefine the LICR command \textperiodcentered to
have appropriate kerning. Finally, when compiling with a non-unicode
engine in a plain TeX setting, we keep the previous definition
(handcrafting the interpunct using a dot).